- Transform bugfixing - Code cleanup - Instancing bugfixing

+++ Engine/Content/Shaders/ShaderCore.hlsl
@@ -9,7 +9,7 @@ struct Scene
 struct SMInstance
 {
     float4x4 World;
-    uint64_t MeshID;
-    uint64_t MaterialID;
+    uint MeshID;
+    uint MaterialID;
     uint MaterialIndex;
 };
